jobengine/requirements.txt
Django==3.1.2
uwsgi==2.0.19.1
python-crontab==2.5.1
